requirements.txt
pillow
google-genai
